+++ migration/b2b/changelog.md
@@ -4,3 +4,11 @@
 
 ### Order 欄位變動
 * Order 新增 token 欄位，避免訂單重複。
+
+## v2
+
+### User 欄位新增
+* User 新增 postCode , city, district
+
+### Message 欄位變動
+* type 欄位減少 orderSync